remonveBindingConstraint compatible v860 + tests

+++ R/removeBindingConstraint.R
@@ -48,6 +48,7 @@ removeBindingConstraint <- function(name, opts = antaresRead::simOptions()) {
   
   namesbc <- unlist(lapply(bindingConstraints, `[[`, "name"), use.names = FALSE)
   
+  # suppression txt files + remove constraint from ini file
   for (i in name) {
     if (! i %in% namesbc) {
       warning(paste0("No binding constraint with name '", i, "'"))
@@ -55,8 +56,24 @@ removeBindingConstraint <- function(name, opts = antaresRead::simOptions()) {
       index <- which(namesbc == i)
       id <- bindingConstraints[[index]]$id
       bindingConstraints[[index]] <- NULL
-      pathValues <- file.path(opts$inputPath, "bindingconstraints", paste0(id, ".txt"))
-      unlink(x = pathValues)
+      # v860
+      if(opts$antaresVersion>=860){
+        path_lt <- file.path(opts$inputPath, 
+                             sprintf("bindingconstraints/%s.txt", 
+                                     paste0(id, "_lt")))
+        path_gt <- file.path(opts$inputPath, 
+                             sprintf("bindingconstraints/%s.txt", 
+                                     paste0(id, "_gt")))
+        path_eq <- file.path(opts$inputPath, 
+                             sprintf("bindingconstraints/%s.txt", 
+                                     paste0(id, "_eq")))
+        lapply(c(path_lt, path_gt, path_eq), 
+               unlink)
+      }else{
+        pathValues <- file.path(opts$inputPath, "bindingconstraints", paste0(id, ".txt"))
+        unlink(x = pathValues)
+      }
+      
       namesbc <- unlist(lapply(bindingConstraints, `[[`, "name"), use.names = FALSE)
       names(bindingConstraints) <- as.character(seq_along(bindingConstraints) - 1)
     }
